requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-snowflake==5.3.0
snowflake-connector-python==3.6.0
Faker==18.9.0
black==24.3.0
flake8==7.0.0
pre-commit==3.7.0
